Added more server-side input checks
Corrected some logic for the sake of cleanliness/consistency.

+++ backend/controllers/authController.js
@@ -23,7 +23,7 @@ module.exports.signup_post = async (req, res) => {
    process.env.NODE_ENV !== "test"
      ? Number(process.env.MAX_USERS)
      : Number(process.env.TEST_MAX_USERS);
- if (registeredUsers.length >= limit) {
+ if (registeredUsers.length === limit) {
    const id = registeredUsers[0]._id;
    await User.findOneAndDelete({ _id: id });
    await Workout.deleteMany({user_id: id});
@@ -49,12 +49,15 @@ module.exports.signup_post = async (req, res) => {
 
 module.exports.verify_user = async (req, res) => {
   const {accountConfirmationToken} = req.params;
+  if(!accountConfirmationToken){
+    res.status(404).json({ error: "Account confirmation token not found"})
+  }
   try {
    const user = await User.findOne({
     accountConfirmationToken: accountConfirmationToken,
   });
   if(!user){
-    return res.status(404).json({error: "User not found."})
+    return res.status(404).json({error: "Couldn't find user with provided confirmation token - this might be because the account has already been confirmed"})
   }
   user.accountStatus = "active";
   user.accountConfirmationToken = undefined;
@@ -82,6 +85,23 @@ module.exports.login_post = async (req, res) => {
 }
 
 module.exports.user_update_patch = async (req, res) => {
+  if(!req.user){
+    res.status(401).json({error: "Not authorized"})
+    return
+  };
+  if(req.body.username && req.body.username.length > 12){
+    res.status(400).json({error: "Too long name."})
+    return
+  }
+  if(req.body.profileImg && 
+     !req.body.profileImg.match(/^data:image\/jpeg/) &&
+     !req.body.profileImg.match(/^data:image\/png/) &&
+     !req.body.profileImg.match(/^data:image\/svg/) 
+     ){
+    res.status(400).json({error: "Bad input, must be JPEG, PNG or SVG image format"})
+    return
+  }
+
   const {id} = req.params;
   const token = req.headers.authorization.slice(7);
   try {
@@ -99,17 +119,21 @@ module.exports.user_update_patch = async (req, res) => {
        success: "Profile updated.",
      });
   } catch (error) {
-    res.status(400).json({ error: "Something went wrong. Please refresh and try again.", logError: error.message });
+    res.status(500).json({ error: "Something went wrong. Please refresh and try again.", logError: error.message });
   }
 };
 
 module.exports.user_deletion = async (req, res) => {
   const { id } = req.params;
+  if(!id){
+      res.status(401).json({error: "Not authorized"})
+    return
+  }
   try {
     const user = await User.findOneAndDelete({ _id: id });
    res.status(200).json({success: "Account deleted successfully"});
   } catch (error) {
-    res.status(400).json({error: "Something went wrong. Please try later.", logError: error.message})
+    res.status(500).json({error: "Something went wrong. Please try later.", logError: error.message})
   }
 };
